feat: Add support to toggle updating of copyright year
Added a function and command to toggle a per-buffer flag to control
updating copyright year.  Modified the update function to check for the
buffer flag before updating the year.

Occasionally I need to edit a file which has a copyright in it but does
not need the date updating - for example a diff or including part of
another file as a quote or example.  This allows the editing to be done
without changing the copyright dates in error.