bug: Add plugin file to set up commands
Moved the plugin command definition into a plugin file from the autoload
file. Renamed the autoload function to toggle the current buffer's
Copyright update flag and documented it. Added a step to the toggle
test file.

If no Copyright owners have been set then the :CopyrightUpdateToggle
command is not known since it is defined in the autoload file.