feat: Prevent putting search pattern in search history
Added the keeppatterns command and changed the original
g#<pat1>#s#<pat2>#<update># command with a %s#<pat>#<update># command.

The pattern used to search for a Copyright notice to update was always
being added to the search history. Adding keeppatterns excluded the
pattern used in the g command but the pattern used to updated the year
would still appear in the search history. Replacing the two commands
with a single file search and replace uses only 1 pattern and
keeppatterns prevents it appearing in the search history.