Add support for single year Copyright statements
Some Copyright owners prefer to just have the latest date in Copyright
statements. Add new commands and functions for updating the Copyright
statement with just the current year.

Added new commands and functions to updating the end year of Copyright
statements with a year range. Deprecated the original command and
function for setting up the Copyright statement update.